requirements.txt
azure-identity
#chromadb
# azure-identity==1.13.0
# azure-search-documents==11.4.0b6
# azure-ai-formrecognizer==3.2.1
azure-storage-blob
langchain
openai==0.28.1
pinecone-client==2.2.2
python-dotenv
sentence-transformers==2.2.2
streamlit
tiktoken
tqdm
unstructured[docx]
unstructured[pdf]
